2016.06.25 - Many improvements, check complete commit
Entities now use components to store and propagate their states, this
works a lot like a proper ECS but isn't really one.

Items are now defined by an ItemType class and the Item class itself is
instancied in every item of the game, allowing more freedom with what
items can do and store. The ItemData thing has been totally removed.

Along with progress on the game engine, many interfaces related to
serialization ( disk as well as networked ) have appeared

PacketSynch are packets that are handled synchronously to the world
ticking

Cleanup of a lot of classes/intefaces, more/better doc

World has methods to place voxels using an entity as parameter, this is
used for voxels that have specific behaviour when placed

Voxels can now have interactions ( when an input is pressed ) and logic
when placed and destroyed ( used for instance in stairs blocks to
determine direction from builder )