Major rework, newer interfaces and better design decisions
## Re-did how voxel data is accessed

Voxel data is now obtained through a suite of poke() / peek() methods,
some making use of improved VoxelContexts, some just being bare and fast
int bit-banging. The fancy voxel updating procedure is now fully in the
hands of the chunk itself, and the world methods are just aliases for
convenience.

## Interface changes for voxels

I added a bunch of world-data related exceptions, thrown when accessing
unloaded bits of the map. In addition to this, many of the voxel
interfaces that were using bare xyz coordinates, now use a single
VoxelContext update as an argument.

## Reworking entities

Entities now use a Location in their constructor, and some stuff has
been pushed to EntityLiving rather than just the bare entity, since it's
irrelevant to most.

Entities are now living inside of chunks, rather than regions.

## Removed EntityVoxel and redid those

EntityVoxel/VoxelEntity was a terrible mess, a hack built quickly. The
entire idea of having entities serve as extra metadata doesn't work any
well for chests and signs, so these were scrapped entirely.

In their place, I'm introducing voxel components, these are stored in a
hashmap in the chunks directly, are named and generally viewed the same
as the voxel data in terms of networking sync.

## Proper way of handling chunk updates

Following the complications due to moving the blunder of syncing
components to the chunks, I decided to finally rethink how chunk uodates
are really done. The registerUser/unregisterUser functions used to
control map loading in singleplayer will now actually be networked, so
the player decides what he wants to load, and will tell the server who
may or may not agree to send the data.

The server will then decide, based on wether or not the player is
subscribed to X chunk, to send entity and voxel updates to them.